+++ backend/tests/baserow/contrib/database/search/test_search_receivers.py
@@ -44,7 +44,7 @@ def test_perm_deleted_workspace(data_fixture):
 
 @pytest.mark.django_db
 @patch(
-    "baserow.contrib.database.search.handler.SearchHandler.schedule_delete_search_data"
+    "baserow.contrib.database.search.handler.SearchHandler.mark_search_data_for_deletion"
 )
 def test_perm_deleted_database(mock, data_fixture):
     """
@@ -54,8 +54,8 @@ def test_perm_deleted_database(mock, data_fixture):
     user = data_fixture.create_user()
     workspace = data_fixture.create_workspace(user=user)
     database = data_fixture.create_database_application(workspace=workspace)
-    table_1 = data_fixture.create_database_table(database=database, name="table_1")
-    table_2 = data_fixture.create_database_table(database=database, name="table_2")
+    data_fixture.create_database_table(database=database, name="table_1", order=1)
+    data_fixture.create_database_table(database=database, name="table_2", order=2)
 
     TrashHandler.trash(user, workspace, database, database)
     TrashHandler.permanently_delete(database)
@@ -70,7 +70,7 @@ def test_perm_deleted_database(mock, data_fixture):
 
 @pytest.mark.django_db
 @patch(
-    "baserow.contrib.database.search.handler.SearchHandler.schedule_delete_search_data"
+    "baserow.contrib.database.search.handler.SearchHandler.mark_search_data_for_deletion"
 )
 def test_perm_deleted_table(mock, data_fixture):
     """
@@ -93,7 +93,7 @@ def test_perm_deleted_table(mock, data_fixture):
 
 @pytest.mark.django_db
 @patch(
-    "baserow.contrib.database.search.handler.SearchHandler.schedule_delete_search_data"
+    "baserow.contrib.database.search.handler.SearchHandler.mark_search_data_for_deletion"
 )
 def test_perm_deleted_field(mock, data_fixture):
     """
@@ -118,7 +118,7 @@ def test_perm_deleted_field(mock, data_fixture):
 
 @pytest.mark.django_db
 @patch(
-    "baserow.contrib.database.search.handler.SearchHandler.schedule_delete_search_data"
+    "baserow.contrib.database.search.handler.SearchHandler.mark_search_data_for_deletion"
 )
 def test_perm_deleted_row(mock, data_fixture):
     """
@@ -144,7 +144,7 @@ def test_perm_deleted_row(mock, data_fixture):
 
 @pytest.mark.django_db
 @patch(
-    "baserow.contrib.database.search.handler.SearchHandler.schedule_delete_search_data"
+    "baserow.contrib.database.search.handler.SearchHandler.mark_search_data_for_deletion"
 )
 def test_perm_deleted_rows(mock, data_fixture):
     """