+++ enterprise/backend/src/baserow_enterprise/apps.py
@@ -246,7 +246,11 @@ class BaserowEnterpriseConfig(AppConfig):
         webhook_event_type_registry.register(RowsEnterViewEventType())
 
         # Create default roles
-        post_migrate.connect(sync_default_roles_after_migrate, sender=self)
+        post_migrate.connect(
+            sync_default_roles_after_migrate,
+            sender=self,
+            dispatch_uid="sync_default_roles_after_migrate",
+        )
 
         from baserow_enterprise.teams.receivers import (
             connect_to_post_delete_signals_to_cascade_deletion_to_team_subjects,
@@ -292,20 +296,48 @@ def sync_default_roles_after_migrate(sender, **kwargs):
             # Note: we used to migrate `NO_ROLE` to `NO_ACCESS` here.
             # This was moved to 0010_rename_no_role_to_no_access.
             with LockedAtomicTransaction(Role):
-                for role_name, operations in tqdm(
+                all_old_roles = {
+                    r.uid: r for r in Role.objects.all().prefetch_related("operations")
+                }
+                all_old_operations = {op.name: op for op in Operation.objects.all()}
+
+                for role_name, role_operations in tqdm(
                     default_roles.items(), desc="Syncing default roles"
                 ):
-                    role, _ = Role.objects.update_or_create(
-                        uid=role_name,
-                        defaults={"name": f"role.{role_name}", "default": True},
-                    )
-                    role.operations.clear()
-
-                    to_add = []
-                    for operation_type in operations:
-                        operation, _ = Operation.objects.get_or_create(
-                            name=operation_type.type
+                    # Create any missing role or update existing ones
+                    role = all_old_roles.get(role_name, None)
+                    if role is None:
+                        role = Role.objects.create(
+                            uid=role_name,
+                            name=f"role.{role_name}",
+                            default=True,
                         )
-                        to_add.append(operation)
+                    elif not role.default or role.name != f"role.{role_name}":
+                        role.name = f"role.{role_name}"
+                        role.default = True
+                        role.save(update_fields=["name", "default"])
 
-                    role.operations.add(*to_add)
+                    # Create any missing operations for the role
+                    new_ops = Operation.objects.bulk_create(
+                        [
+                            Operation(name=op.type)
+                            for op in role_operations
+                            if op.type not in all_old_operations
+                        ],
+                    )
+                    all_old_operations.update({op.name: op for op in new_ops})
+
+                    old_role_ops = set(op.name for op in role.operations.all())
+                    new_role_ops = set(op.type for op in role_operations)
+
+                    roles_to_add = new_role_ops - old_role_ops
+                    if roles_to_add:
+                        role.operations.add(
+                            *[all_old_operations[op] for op in roles_to_add],
+                        )
+
+                    to_remove = old_role_ops - new_role_ops
+                    if to_remove:
+                        role.operations.remove(
+                            *[all_old_operations[op] for op in to_remove],
+                        )