+++ web-frontend/modules/integrations/localBaserow/components/services/LocalBaserowAggregateRowsForm.vue
@@ -55,57 +55,28 @@
         </FormGroup>
       </div>
     </div>
-    <div v-if="!small && !fieldsLoading" class="row">
-      <div class="col col-12">
-        <Tabs>
-          <Tab
-            :title="$t('localBaserowAggregateRowsForm.filterTabTitle')"
-            class="service-form__condition-form-tab"
-          >
-            <LocalBaserowTableServiceConditionalForm
-              v-if="values.table_id"
-              v-model="values.filters"
-              :fields="tableFields"
-              :filter-type.sync="values.filter_type"
-            >
-            </LocalBaserowTableServiceConditionalForm>
-            <p v-if="!values.table_id">
-              {{
-                $t('localBaserowAggregateRowsForm.noTableChosenForFiltering')
-              }}
-            </p>
-          </Tab>
-          <Tab
-            :title="$t('localBaserowAggregateRowsForm.searchTabTitle')"
-            class="service-form__search-form-tab"
-          >
-            <InjectedFormulaInput
-              v-model="values.search_query"
-              small
-              :placeholder="
-                $t('localBaserowAggregateRowsForm.searchFieldPlaceHolder')
-              "
-            />
-          </Tab>
-        </Tabs>
-      </div>
-    </div>
+    <ServiceRefinementForms
+      v-if="!fieldsLoading && values.table_id"
+      :small="small"
+      :values="values"
+      :table-fields="tableFields"
+      show-filter
+      show-search
+    />
     <div v-if="fieldsLoading" class="loading-spinner"></div>
   </form>
 </template>
 
 <script>
 import form from '@baserow/modules/core/mixins/form'
-import LocalBaserowTableServiceConditionalForm from '@baserow/modules/integrations/localBaserow/components/services/LocalBaserowTableServiceConditionalForm'
-import InjectedFormulaInput from '@baserow/modules/core/components/formula/InjectedFormulaInput'
 import LocalBaserowServiceForm from '@baserow/modules/integrations/localBaserow/components/services/LocalBaserowServiceForm.vue'
 import localBaserowService from '@baserow/modules/integrations/localBaserow/mixins/localBaserowService'
+import ServiceRefinementForms from '@baserow/modules/integrations/localBaserow/components/services/ServiceRefinementForms'
 
 export default {
   components: {
     LocalBaserowServiceForm,
-    InjectedFormulaInput,
-    LocalBaserowTableServiceConditionalForm,
+    ServiceRefinementForms,
   },
   mixins: [form, localBaserowService],
   data() {