+++ web-frontend/modules/integrations/localBaserow/components/services/LocalBaserowSignalTriggerServiceForm.vue
@@ -3,6 +3,7 @@
     <LocalBaserowServiceForm
       :application="application"
       :default-values="defaultValues"
+      :enable-view-picker="false"
       @values-changed="emitServiceChange($event)"
     ></LocalBaserowServiceForm>
   </form>